+++ server-nodejs/db/mongodb.js
@@ -14,7 +14,6 @@ module.exports = function (conf) {
 
     var mongo = require('mongodb');
     var ObjectID = mongo.ObjectID;
-    require('./utils');
 
     /*
      * Initialize the connection.
