+++ server-nodejs/index.js
@@ -11,9 +11,6 @@ var debug = require('debug')('app:' + process.pid);
 debug('Initializing express');
 var express = require('express');
 var app = express();
-var fs = require('fs');
-var socket = require('./events/socket');
-
 
 /*
  * Configuration
@@ -22,9 +19,6 @@ debug('Loading configuration');
 
 var conf = app.locals.conf = require('./conf');
 app.locals.db = require('./db')(conf.db, conf[conf.db]);
-
-var http_port = process.env.HTTP_PORT || 8090;
-var https_port = process.env.HTTPS_PORT || 8443;
 
 require('./routes')(app, express);
 
@@ -39,28 +33,33 @@ if (module.parent) {
 
 debug('Working in %s mode', app.get('env'));
 
-/*
- * Create an HTTP server
- */
-debug('Creating HTTP server on port %s', http_port);
-var http = require('http').createServer(app).listen(http_port, function () {
-    debug('HTTP server listening on port %s', http_port);
-    socket.attach(http);
-});
+function startServer(type, port) {
+    global.server = server;
+    server.listen(port, function () {
+        debug('%s server listening on port %s', type, port);
+        var socket = require('./events/socket');
+        socket.attach(server);
+    });
+}
 
 /*
  * Create a HTTPS server if there are certificates
+ * Create a HTTP server otherwise
  */
-if (conf.connection.hasOwnProperty('Key') &&
-        conf.connection.hasOwnProperty('Cert')) {
-    debug('Creating HTTPS server on port %s', https_port);
-    var https = require('https').createServer({
-        key : fs.readFileSync(conf.connection.Key).toString(),
-        cert : fs.readFileSync(conf.connection.Cert).toString()
-    }, app).listen(https_port, function () {
-        debug('HTTPS server listening on port %s', https_port);
-        socket.attach(https);
-    });
+if (conf.connection && conf.connection.Key && conf.connection.Cert) {
+    var fs = require('fs');
+    var port = process.env.HTTPS_PORT || 8443;
+    debug('Creating HTTPS server on port %s', port);
+    global.server = require('https').createServer({
+        key: fs.readFileSync(conf.connection.Key).toString(),
+        cert: fs.readFileSync(conf.connection.Cert).toString()
+    }, app);
+    startServer('HTTPS', port);
+} else {
+    var port = process.env.HTTP_PORT || 8090;
+    debug('Creating HTTP server on port %s', port);
+    global.server = require('http').createServer(app);
+    startServer('HTTPS', port);
 }
